Re-use renderRuleEditor method
