+++ public/app/features/alerting/unified/RuleEditorCloudOnlyAllowed.test.tsx
@@ -1,20 +1,15 @@
-import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
+import { screen, waitForElementToBeRemoved } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
-import { Provider } from 'react-redux';
-import { Route, Router } from 'react-router-dom';
-import { ui } from 'test/helpers/alertingRuleEditor';
-import { byRole, byTestId, byText } from 'testing-library-selector';
+import { renderRuleEditor, ui } from 'test/helpers/alertingRuleEditor';
+import { byRole, byText } from 'testing-library-selector';
 
-import { selectors } from '@grafana/e2e-selectors';
-import { locationService, setDataSourceSrv } from '@grafana/runtime';
+import { setDataSourceSrv } from '@grafana/runtime';
 import { contextSrv } from 'app/core/services/context_srv';
-import { configureStore } from 'app/store/configureStore';
 import { PromApplication } from 'app/types/unified-alerting-dto';
 
 import { searchFolders } from '../../manage-dashboards/state/actions';
 
-import RuleEditor from './RuleEditor';
 import { discoverFeatures } from './api/buildInfo';
 import { fetchRulerRules, fetchRulerRulesGroup, fetchRulerRulesNamespace, setRulerRuleGroup } from './api/ruler';
 import { ExpressionEditorProps } from './components/rule-editor/ExpressionEditor';
@@ -55,20 +50,6 @@ const mocks = {
     fetchRulerRulesIfNotFetchedYet: jest.mocked(fetchRulerRulesIfNotFetchedYet),
   },
 };
-
-function renderRuleEditor(identifier?: string) {
-  const store = configureStore();
-
-  locationService.push(identifier ? `/alerting/${identifier}/edit` : `/alerting/new`);
-
-  return render(
-    <Provider store={store}>
-      <Router history={locationService.getHistory()}>
-        <Route path={['/alerting/new', '/alerting/:id/edit']} component={RuleEditor} />
-      </Router>
-    </Provider>
-  );
-}
 
 describe('RuleEditor cloud: checking editable data sources', () => {
   beforeEach(() => {
